+++ src/main/kotlin/dev/ekvedaras/laravelquery/models/SettingsSchema.kt
@@ -8,8 +8,13 @@ class SettingsSchema(
     val schema: DasNamespace
 ) {
     companion object {
-        const val separator = "<-/lq/->"
+        private const val separator = "<-/lq/->"
+        fun keyFor(schemaName: String, dataSourceId: String) = dataSourceId + separator + schemaName
+
+        fun keyFor(dasNamespace: DasNamespace, dbDataSource: DbDataSource) =
+            keyFor(dasNamespace.name, dbDataSource.uniqueId)
     }
+
     override fun toString(): String = schema.name
-    fun key(): String = "${dataSource.uniqueId}$separator${schema.name}"
+    fun key(): String = keyFor(schema, dataSource)
 }